requirements.txt
Django
psycopg2-binary
celery[redis]
django-celery-beat
djangorestframework
django-filter
drf-spectacular
drf-spectacular[sidecar]
pytest
pytest-django
pytest-factoryboy
django-rest-framework-nested
djangorestframework-simplejwt
django-cors-headers
pillow
boto3
openpyxl
reportlab
drf-nested-routers
django-extensions